requirements.txt
spotipy
ytmusicapi
python-dotenv
